requirements.txt
pycryptodome
sympy
matplotlib
numpy
pytest